03_web_search/requirements.txt
dokis
ollama
requests
beautifulsoup4
python-dotenv
